Add files via upload
Build lemonade stand budget tracker using a Python application that stores data in a .csv file 
•	The program allows you to track multiple accounts.
•	The program can analyze profit and loss for a specific month or year
•	The program can provide high level stats, such as average transaction size.
•	The program can track and analyze transactions by specific customers or vendors.
Budgeter
	User Stories
1.	When starting up the application, following options are given :
a.	Add a new entry to the budget tracker.
b.	Display the total account balance.
c.	View all previous entries.
2.	If a new entry is added,  code ask to provide:
a.	A title describing the budget item.
b.	Whether the budget item is Income or Expense.
c.	The total amount of the budget item.
d.	The date of the transaction in "MM-DD-YYYY" string format.
3.	For displaying the total account balance:
a.	The program adds all income and subtracts all expense items to display the net balance.
4.	For viewing all previous entries:
a.	The program prints all details of all previous entries in a human readable format.
Technical Requirements
•	Stores all entries in a .csv file.
•	Load the previously created entries when the user initializes the application.